Separated Script for AI
