+++ Assets/Scripts/Enemies/Melee/MeleeAI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c7ad9c663ce7d084fa5e197259281033
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
